+++ CHANGELOG.md
@@ -3,7 +3,19 @@
 > [!NOTE]
 > You find a complete Changelog here [.changelog/Changelog.md](https://github.com/cheinisch/minniark/blob/main/.changelog/Changelog.md)
 
-## [2025.12.0]
+## [2025.12.2]
+2025-12-15
+
+**Full Changelog**: https://github.com/cheinisch/minniark/compare/v2025.11.2...v2025.12.2
+
+### Changes
+- Album get new Design
+- Collection get new Design
+
+### Bugfixes
+- add language missing language strings
+
+## [2025.12.1]
 2025-12-14
 
 **Full Changelog**: https://github.com/cheinisch/minniark/compare/v2025.11.2...v2025.12.1